--- starter_code/app.js
@@ -1,18 +1,0 @@
-
-const express = require('express');
-const hbs     = require('hbs');
-const app     = express();
-const path    = require('path');
-const PunkAPIWrapper = require('punkapi-javascript-wrapper');
-const punkAPI = new PunkAPIWrapper();
-
-app.set('view engine', 'hbs');
-app.set('views', __dirname + '/views');
-app.use(express.static(path.join(__dirname, 'public')));
-
-app.get('/', (req, res, next) => {
-  res.render('index');
-});
-
-
-app.listen(3000);
