--- starter_code/views/index.hbs
@@ -1,3 +1,0 @@
-<div class="index-container">
-  
-</div>
