+++ views/beers.hbs
@@ -1,0 +1,23 @@
+<!-- beers.hbs -->
+{{!-- <p>tertertertertertter</p> --}}
+
+<div class="container-beers-list">
+
+{{#each this}} 
+ 
+   {{> beersPartial }}
+{{/each}}
+
+</div>
+
+
+{{!-- Aixi funciona per imprimir els valors --}}
+{{!-- <div>
+<ul>
+{{#each this}} 
+  <li>{{this.name}}</li> --}}
+   {{!-- {{> beersPartial }}
+{{/each}}
+</ul>
+</div> --}}
+
