+++ views/index.hbs
@@ -1,0 +1,8 @@
+<!-- index.hbs -->
+<div class="index-container">
+  
+    <div class="container-beer"><img class="image-beer" src="images/beer.png"></div>
+    <div class="button-beers"><input type="button" class="button" onclick="location.href='/beers';" value="Check the beers!"></div>
+    <div class="button-beers"><input type="button" class="button" onclick="location.href='/random-beers';" value="Check a Random beer!"></div>
+  
+</div>
