+++ views/layout.hbs
@@ -1,0 +1,23 @@
+<!-- layout.hbs -->
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0"> 
+  <title>IronBeers</title>
+  <link rel="stylesheet" href="stylesheets/normalize.css">
+  <link rel="stylesheet" href="stylesheets/styles.css">
+</head>
+<body>
+  <nav class="navbar">
+    <div><a class="text-navbar" href="/">Home</a></div>
+    <div><a class="text-navbar" href="/beers">Beers</a></div>
+    <div><a class="text-navbar" href="/random-beers">Random Beer</a></div>
+  </nav>
+  {{{ body }}}
+  <div class="footer">
+    <p>Develop by Ironhack ;)</p>
+  </div>
+</body>
+</html>
