+++ views/partials/beersPartial.hbs
@@ -1,0 +1,8 @@
+<div class="individual-beer">
+  <div><img src={{this.image_url}}></div>
+  <div class="individual-beer-name">{{this.name}}</div>
+  <div class="individual-beer-tag">{{this.tagline}}</div>
+  <div class="individual-beer-description">{{this.description}}</div>
+  
+
+</div>
