+++ views/randomBeer.hbs
@@ -1,0 +1,9 @@
+<div class="container-random-beer">
+  <p>hola que tal</p>
+
+
+ 
+   {{> beersPartial this }}
+
+
+</div>
